+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+# 0.2.0
+- Removed pattern search functionality.
+- `gothicAudio.audioDir` is not set by default now.
+- `gothicAudio.audioDir` will show an popup allowing to choose the Gothic audio directory.
+- `gothic-audio-player.fixSilesian` is now optional and isn't registered by default.
+
 # 0.1.3
 - Added functionality to work with  `SVM` files.